renamed conference to defaultProps type
